requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
firebase-admin
pydantic
python-dotenv
